Refactor FindACourseApi models and features
