--- src/Dfc.CourseDirectory.FindACourseApi/DTOs/Coordinates.cs
@@ -1,8 +1,0 @@
-﻿namespace Dfc.CourseDirectory.FindACourseApi.DTOs
-{
-    public class Coordinates
-    {
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
-    }
-}
