--- src/Dfc.CourseDirectory.FindACourseApi/DTOs/FacetCountResult.cs
@@ -1,8 +1,0 @@
-﻿namespace Dfc.CourseDirectory.FindACourseApi.DTOs
-{
-    public class FacetCountResult
-    {
-        public long Count { get; set; }
-        public object Value { get; set; }
-    }
-}
