+++ src/Dfc.CourseDirectory.FindACourseApi/Features/CourseRunDetail/QualificationViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Dfc.CourseDirectory.FindACourseApi.ViewModels
+﻿namespace Dfc.CourseDirectory.FindACourseApi.Features.CourseRunDetail
 {
     public class QualificationViewModel
     {
