+++ src/Dfc.CourseDirectory.FindACourseApi/Features/CourseSearch/CoordinatesViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Dfc.CourseDirectory.FindACourseApi.Features.CourseSearch
+{
+    public class CoordinatesViewModel
+    {
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
+    }
+}
