+++ src/Dfc.CourseDirectory.FindACourseApi/Features/CourseSearch/FacetCountResultViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Dfc.CourseDirectory.FindACourseApi.Features.CourseSearch
+{
+    public class FacetCountResultViewModel
+    {
+        public long Count { get; set; }
+        public object Value { get; set; }
+    }
+}
